refactor(data_structures): decouple linear and linked list demos to demos/data_structures
